+++ CHANGES.rst
@@ -11,6 +11,7 @@ in progress
 - Add "geohash" field when writing into InfluxDB
 - Use nominatim.hiveeyes.org as primary reverse geocoder,
   fall back to nominatim.openstreetmap.org
+- Add option to disable the Nominatim cache
 
 
 2018-12-11 0.4.3